backend/pyproject.toml
[tool.poetry]
name = "blog"
version = "0.1.0"
description = ""
authors = ["Ali Elbashir <alielbashiir@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.75.1"
uvicorn = "^0.17.6"
beanie = "^1.10.8"
python-decouple = "^3.6"
PyJWT = "^2.3.0"
passlib = "^1.7.4"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
requests = "^2.27.1"
black = "^22.3.0"
pre-commit = "^2.18.1"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
